+++ modules/Vector_Operations.cpp
@@ -12,7 +12,7 @@ Vector operator+(const Vector &v, const Vector &u) {
     if (!equal_size(v, u))
         return new_vector;
 
-    for (int i = 1; i <= v.size(); i++)
+    for (size_t i = 1; i <= v.size(); i++)
         new_vector(i) = v.elem(i) + u.elem(i);
 
     return new_vector; 
@@ -23,7 +23,7 @@ Vector operator-(const Vector &v, const Vector &u) {
     if (!equal_size(v, u))
         return new_vector;
 
-    for (int i = 1; i <= v.size(); i++)
+    for (size_t i = 1; i <= v.size(); i++)
         new_vector(i) = v.elem(i) - u.elem(i);
 
     return new_vector; 
@@ -32,7 +32,7 @@ Vector operator-(const Vector &v, const Vector &u) {
 Vector operator*(double k, const Vector &v) {
     Vector new_vector(v.size());
 
-    for (int i = 1; i <= v.size(); i++)
+    for (size_t i = 1; i <= v.size(); i++)
         new_vector(i) = k * v.elem(i);
   
     return new_vector;
@@ -42,7 +42,7 @@ bool operator==(const Vector &v, const Vector &u) {
     if (!equal_size(v, u))
         return false;
     
-    for (int i = 1; i <= v.size(); i++)
+    for (size_t i = 1; i <= v.size(); i++)
         if (v.elem(i) != u.elem(i))
             return false;
 
@@ -59,7 +59,7 @@ bool operator!=(const Vector &v, const Vector &u) {
     if (!equal_size(v, u))
         return true;
 
-    for (int i = 0; i <= v.size(); i++)
+    for (size_t i = 0; i <= v.size(); i++)
         if (v.elem(i) == u.elem(i))
             return false;
 
@@ -77,7 +77,7 @@ double dot_product(const Vector &v, const Vector &u) {
     if (!equal_size(v, u))
         return sum;
 
-    for (int i = 0; i <= v.size(); i++)
+    for (size_t i = 0; i <= v.size(); i++)
         sum += v.elem(i) * u.elem(i);
 
     return sum;
@@ -89,13 +89,13 @@ Vector cross_product(const Vector &u, const Vector &v) {
         return new_vector;
 
     Matrix matrix(3, u.size());
-    for (int i = 1; i <= u.size(); i++) {
+    for (size_t i = 1; i <= u.size(); i++) {
         matrix(1, i) = 1;
         matrix(2, i) = u.elem(i);
         matrix(3, i) = v.elem(i);
     }
 
-    for (int i = 1; i <= u.size(); i ++)
+    for (size_t i = 1; i <= u.size(); i ++)
         new_vector(i) = det(sub_matrix(matrix, i));
     
     return new_vector;
@@ -104,7 +104,7 @@ Vector cross_product(const Vector &u, const Vector &v) {
 Vector tensor_product(const Vector &v, const Vector &u) {
     Vector new_vector(v.size() * u.size());
 
-    for (int i = 1, j = 1, k = 1; i <= new_vector.size(); i++, k++) {
+    for (size_t i = 1, j = 1, k = 1; i <= new_vector.size(); i++, k++) {
         if (k > u.size()) {
             k = 1;
             j++;
@@ -119,7 +119,7 @@ Vector tensor_product(const Vector &v, const Vector &u) {
 double vector_length(const Vector &vector) {
     double sum = 0.0;
 
-    for (int i = 1; i <= vector.size(); i++)
+    for (size_t i = 1; i <= vector.size(); i++)
         sum += vector.elem(i) * vector.elem(i);
 
     return sqrt(sum);
